requirements.txt
# Google API client libraries
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# OpenAI for LLM integration
openai==1.23.2

# Slack SDK
slack-sdk==3.27.1

# Utilities
python-dotenv==1.0.1
requests==2.31.0
pytz==2024.1

# For SQLite (built into Python, no external library needed)

# Optional: for regex and parsing (already built-in, so no need to install separately)
# re, base64, datetime, os, sys, json, email.utils
